ReOrganized the ProjectManager and Project Controller files
