+++ herckules/src/main/java/edu/asu/diging/lerna/herckules/web/ProjectController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import edu.asu.diging.lerna.herckules.authentication.IUser;
 import edu.asu.diging.lerna.herckules.authentication.IUserManager;
-import edu.asu.diging.lerna.herckules.authentication.impl.IProjectManager;
+import edu.asu.diging.lerna.herckules.db.IProjectManager;
 import edu.asu.diging.lerna.herckules.domain.IProject;
 import edu.asu.diging.lerna.herckules.domain.impl.Project;
